requirements.txt
playwright
pytest-playwright
pytest
pytest-html
pytest-xdist
PyYAML
python-dotenv
allure-pytest
extend-report-selenium